feat: add migrations, add db

+++ RazorPagesProject/Pages/Employees/Edit.cshtml.cs
@@ -52,7 +52,7 @@ namespace RazorPagesProject.Pages.Employees
                     if (Employee.PhotoPath != null)
                     {
                         string filePath = Path.Combine(_webHostEnvironment.WebRootPath, "images", Employee.PhotoPath);
-                        System.IO.File.Delete(filePath);
+                        if (Employee.PhotoPath != "noimage.png") System.IO.File.Delete(filePath);
                     }
 
                     Employee.PhotoPath = processUploadedFile();
